agentic-day1/requirements.txt
langchain
langchain-openai
langchain-core
python-dotenv
openai